Done: get lativ styles and rename style name
